changed datatype to float

+++ svm.h
@@ -42,18 +42,18 @@ struct svm_parameter
 	int svm_type;
 	int kernel_type;
 	int degree;	/* for poly */
-	double gamma;	/* for poly/rbf/sigmoid */
-	double coef0;	/* for poly/sigmoid */
+	float gamma;	/* for poly/rbf/sigmoid */
+	float coef0;	/* for poly/sigmoid */
 
 	/* these are for training only */
-	double cache_size; /* in MB */
-	double eps;	/* stopping criteria */
-	double C;	/* for C_SVC, EPSILON_SVR and NU_SVR */
+	float cache_size; /* in MB */
+	float eps;	/* stopping criteria */
+	float C;	/* for C_SVC, EPSILON_SVR and NU_SVR */
 	int nr_weight;		/* for C_SVC */
 	int *weight_label;	/* for C_SVC */
-	double* weight;		/* for C_SVC */
-	double nu;	/* for NU_SVC, ONE_CLASS, and NU_SVR */
-	double p;	/* for EPSILON_SVR */
+	float* weight;		/* for C_SVC */
+	float nu;	/* for NU_SVC, ONE_CLASS, and NU_SVR */
+	float p;	/* for EPSILON_SVR */
 	int shrinking;	/* use the shrinking heuristics */
 	int probability; /* do probability estimates */
 };
@@ -67,11 +67,11 @@ struct svm_model
 	int nr_class;		/* number of classes, = 2 in regression/one class svm */
 	int l;			/* total #SV */
 	struct svm_node **SV;		/* SVs (SV[l]) */
-	double **sv_coef;	/* coefficients for SVs in decision functions (sv_coef[k-1][l]) */
-	double *rho;		/* constants in decision functions (rho[k*(k-1)/2]) */
-	double *probA;		/* pariwise probability information */
-	double *probB;
-	double *prob_density_marks;	/* probability information for ONE_CLASS */
+	float **sv_coef;	/* coefficients for SVs in decision functions (sv_coef[k-1][l]) */
+	float *rho;		/* constants in decision functions (rho[k*(k-1)/2]) */
+	float *probA;		/* pariwise probability information */
+	float *probB;
+	float *prob_density_marks;	/* probability information for ONE_CLASS */
 	int *sv_indices;        /* sv_indices[0,...,nSV-1] are values in [1,...,num_traning_data] to indicate SVs in the training set */
 
 	/* for classification only */
@@ -85,7 +85,7 @@ struct svm_model
 };
 
 struct svm_model *svm_train(const struct svm_problem *prob, const struct svm_parameter *param);
-void svm_cross_validation(const struct svm_problem *prob, const struct svm_parameter *param, int nr_fold, double *target);
+void svm_cross_validation(const struct svm_problem *prob, const struct svm_parameter *param, int nr_fold, float *target);
 
 int svm_save_model(const char *model_file_name, const struct svm_model *model);
 struct svm_model *svm_load_model(const char *model_file_name);
@@ -95,11 +95,11 @@ int svm_get_nr_class(const struct svm_model *model);
 void svm_get_labels(const struct svm_model *model, int *label);
 void svm_get_sv_indices(const struct svm_model *model, int *sv_indices);
 int svm_get_nr_sv(const struct svm_model *model);
-double svm_get_svr_probability(const struct svm_model *model);
+float svm_get_svr_probability(const struct svm_model *model);
 
-double svm_predict_values(const struct svm_model *model, const struct svm_node *x, double* dec_values);
-double svm_predict(const struct svm_model *model, const struct svm_node *x);
-double svm_predict_probability(const struct svm_model *model, const struct svm_node *x, double* prob_estimates);
+float svm_predict_values(const struct svm_model *model, const struct svm_node *x, float* dec_values);
+float svm_predict(const struct svm_model *model, const struct svm_node *x);
+float svm_predict_probability(const struct svm_model *model, const struct svm_node *x, float* prob_estimates);
 
 void svm_free_model_content(struct svm_model *model_ptr);
 void svm_free_and_destroy_model(struct svm_model **model_ptr_ptr);
